[BE] feat: propertyId Long -> Integer 형변환 (S14P21C103-240)

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/service/game/realestate/response/RealEstateDocumentResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 public record RealEstateDocumentResponse(
-    Long propertyId,
+    Integer propertyId,
     String propertyName,
     String address,
     BigDecimal latitude,
@@ -22,7 +22,7 @@ public record RealEstateDocumentResponse(
     }
 
     public static RealEstateDocumentResponse of(
-        final Long propertyId,
+        final Integer propertyId,
         final String propertyName,
         final String address,
         final BigDecimal latitude,
