+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/service/game/start/response/TargetPropertyListResponse.java
@@ -27,7 +27,7 @@ public class TargetPropertyListResponse {
     @Getter
     public static class TargetPropertyResponse {
 
-        private final Long propertyId;
+        private final Integer propertyId;
         private final String name;
         private final long recentPrice;
         private final BigDecimal latitude;
@@ -35,7 +35,7 @@ public class TargetPropertyListResponse {
 
         @Builder
         private TargetPropertyResponse(
-            final Long propertyId,
+            final Integer propertyId,
             final String name,
             final long recentPrice,
             final BigDecimal latitude,
@@ -52,7 +52,7 @@ public class TargetPropertyListResponse {
         }
 
         public static TargetPropertyResponse of(
-            final Long propertyId,
+            final Integer propertyId,
             final String name,
             final long recentPrice,
             final BigDecimal latitude,
@@ -74,7 +74,7 @@ public class TargetPropertyListResponse {
         }
     }
 
-    private static void validatePropertyId(final Long propertyId) {
+    private static void validatePropertyId(final Integer propertyId) {
         if (propertyId == null) {
             throw new HomerunException(ErrorCode.GLOBAL_CONFIGURATION_INVALID);
         }
